backend/requirements.txt
autopep8==1.6.0
mysqlclient==2.0.3
pycodestyle==2.8.0
python-dotenv==0.19.1
toml==0.10.2

bcrypt==3.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
openpyxl==3.1.1
psycopg2==2.9.5
pycparser==2.21
pycryptodome==3.10.1
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
urllib3==1.26.14
uuid==1.30
Werkzeug==2.2.3
zipp==3.15.0
